+++ packages/xr/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @threlte/xr
 
+## 0.0.4
+
+### Patch Changes
+
+- 96257d4c: Tighten types of `<Controller>`
+- 7498fb62: Added slot `"fallback"` to `<XR>`
+- e71b6408: Add `<Headset />` component to allow attaching to viewer headset.
+- 271a8723: Changed `controller` to `targetRay` (adhering to type and slot name rather than function call)
+
 ## 0.0.3
 
 ### Patch Changes